XorshopWebParent/XorshopBackEnd/src/main/java/com/xorshop/admin/country/state/StateRepository.java
package com.xorshop.admin.country.state;

import java.util.List;

import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.CrudRepository;

import com.xorshop.common.entity.Country;
import com.xorshop.common.entity.State;
public interface StateRepository extends CrudRepository<State, Integer> {

	public List<State> findByCountryOrderByNameAsc(Country country);
	
	@Query("SELECT s FROM State s LEFT JOIN s.country ON s.country.id = s.id WHERE s.name = :name")
	public State findByName(String name);
}
